pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "umbrellaAI"
version = "0.1.0"
description = "A multi-agent AI system for document processing and analysis"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = ["ai", "nlp", "document-processing"]
authors = [
  { name = "Your Name", email = "your.email@example.com" }
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
]
dependencies = [
  "fastapi>=0.100.0",
  "uvicorn>=0.22.0",
  "pydantic>=2.0.0",
  "google-cloud-aiplatform>=1.25.0",
  "langchain>=0.0.200",
  "python-multipart>=0.0.6",
  "aiohttp>=3.8.0",
  "httpx>=0.24.0",
  "python-json-logger>=2.0.0"
]

[project.optional-dependencies]
dev = [
  "black>=22.3.0",
  "isort>=5.10.1",
  "flake8>=4.0.1",
  "mypy>=0.950"
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"] 